+++ src/adapters/AssetsField.php
@@ -94,8 +94,8 @@ class AssetsField
 				}
 
 				$event->controls = Craft::$app->getView()->renderTemplate('blockonomicon/_adapters/AssetsFieldAdapter.html', [
-					'safeHandle' => implode('_', preg_split('/[\[\]]+/', $event->handle, -1, PREG_SPLIT_NO_EMPTY)),
-					'blockHandle' => $event->handle,
+					'safeHandle' => $event->blockHandle . '_' . implode('_', preg_split('/[\[\]]+/', $event->handle, -1, PREG_SPLIT_NO_EMPTY)),
+					'fieldHandle' => $event->handle,
 					'settings' => $event->settings,
 					'cachedOptions' => $event->cachedoptions,
 					'sourceOptions' => $sourceoptions,